feat: seperate api validator to files
