+++ core/api-server/lib/service/builds.js
@@ -19,7 +19,7 @@ const minimumBytes = 4100;
 
 class Builds {
     async getBuild(options) {
-        validator.validateBuildId(options);
+        validator.builds.validateBuildId(options);
         const response = await stateManager.algorithms.builds.get(options);
         if (!response) {
             throw new ResourceNotFoundError('build', options.buildId);
@@ -28,7 +28,7 @@ class Builds {
     }
 
     async getBuilds(options) {
-        validator.validateResultList(options);
+        validator.lists.validateResultList(options);
         const response = await stateManager.algorithms.builds.list(options);
         return response;
     }
@@ -48,7 +48,7 @@ class Builds {
     }
 
     async stopBuild(options) {
-        validator.validateBuildId(options);
+        validator.builds.validateBuildId(options);
         const { buildId } = options;
         const build = await this.getBuild({ buildId });
         if (!this.isActiveState(build.status)) {
@@ -63,7 +63,7 @@ class Builds {
     }
 
     async rerunBuild(options) {
-        validator.validateBuildId(options);
+        validator.builds.validateBuildId(options);
         const { buildId } = options;
         const build = await this.getBuild({ buildId });
         if (this.isActiveState(build.status)) {
@@ -104,7 +104,7 @@ class Builds {
             const version = newAlgorithm.gitRepository.commit.id;
             buildId = this._createBuildID(newAlgorithm.name);
             const { env, name, gitRepository, type, baseImage } = newAlgorithm;
-            validator.validateAlgorithmBuildFromGit({ env });
+            validator.builds.validateAlgorithmBuildFromGit({ env });
             await this.startBuild({ buildId, version, env, algorithmName: name, gitRepository, type, baseImage });
         }
         return { buildId, messages };
@@ -113,7 +113,7 @@ class Builds {
     async _newAlgorithm(file, oldAlgorithm, newAlgorithm) {
         const fileInfo = await this._fileInfo(file);
         const env = this._resolveEnv(oldAlgorithm, newAlgorithm);
-        validator.validateAlgorithmBuild({ fileExt: fileInfo.fileExt, env });
+        validator.builds.validateAlgorithmBuild({ fileExt: fileInfo.fileExt, env });
         return { ...newAlgorithm, fileInfo, env };
     }
