+++ core/api-server/lib/validation/cron.js
@@ -1,0 +1,11 @@
+class ApiValidator {
+    constructor(validator) {
+        this._validator = validator;
+    }
+
+    validateCronRequest(options) {
+        this._validator.validate(this._validator.definitions.cronRequest, options, false);
+    }
+}
+
+module.exports = ApiValidator;
