+++ core/api-server/lib/validation/graphs.js
@@ -1,0 +1,11 @@
+class ApiValidator {
+    constructor(validator) {
+        this._validator = validator;
+    }
+
+    validateGraphQuery(options) {
+        this._validator.validate(this._validator.definitionsInternal.graph, options, true);
+    }
+}
+
+module.exports = ApiValidator;
