+++ core/api-server/lib/validation/index.js
@@ -1,0 +1,25 @@
+const Algorithms = require('./algorithms');
+const Boards = require('./boards');
+const Builds = require('./builds');
+const Cron = require('./cron');
+const Executions = require('./executions');
+const Experiments = require('./experiments');
+const Graphs = require('./graphs');
+const Internal = require('./internal');
+const Jobs = require('./jobs');
+const Lists = require('./lists');
+const Pipelines = require('./pipelines');
+
+module.exports = {
+    Algorithms,
+    Boards,
+    Builds,
+    Cron,
+    Executions,
+    Experiments,
+    Graphs,
+    Internal,
+    Jobs,
+    Lists,
+    Pipelines
+};
