+++ core/api-server/lib/validation/jobs.js
@@ -1,0 +1,15 @@
+class ApiValidator {
+    constructor(validator) {
+        this._validator = validator;
+    }
+
+    validateName(pipeline) {
+        this._validator.validate(this._validator.definitionsInternal.name, pipeline, false);
+    }
+
+    validateJobID(pipeline) {
+        this._validator.validate(this._validator.definitions.jobIdObject, pipeline, false);
+    }
+}
+
+module.exports = ApiValidator;
